update initializeTimes and updateTimes functions to use api functions and fetch data, currently unit tests failing, need to correct that
