+++ public/index.html
@@ -25,6 +25,8 @@
       crossorigin="anonymous"
     ></script>
 
+    <script src="../src/api.js" defer></script>
+
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
 
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
@@ -36,8 +38,6 @@
     <title>Little Lemon</title>
   </head>
   <body>
-    <script src="https://raw.githubusercontent.com/Meta-Front-End-Developer-PC/capstone/master/api.js"></script>
-
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
   </body>
